test: re-organize test case
