+++ tap-snapshots/test-mode-static.js-TAP.test.js
@@ -5,7 +5,7 @@
  * Make sure to inspect the output below.  Do not ignore changes!
  */
 'use strict'
-exports[`test/static.js TAP postProcessor works, swagger route returns updated yaml > must match snapshot 1`] = `
+exports[`test/mode/static.js TAP postProcessor works, swagger route returns updated yaml > must match snapshot 1`] = `
 openapi: 3.0.0
 info:
   description: Test swagger specification
@@ -40,31 +40,31 @@ paths:
 
 `
 
-exports[`test/static.js TAP specification validation check works > must match snapshot 1`] = `
+exports[`test/mode/static.js TAP specification validation check works > must match snapshot 1`] = `
 Error: specification is missing in the module options
 `
 
-exports[`test/static.js TAP specification validation check works > must match snapshot 2`] = `
+exports[`test/mode/static.js TAP specification validation check works > must match snapshot 2`] = `
 Error: specification is not an object
 `
 
-exports[`test/static.js TAP specification validation check works > must match snapshot 3`] = `
+exports[`test/mode/static.js TAP specification validation check works > must match snapshot 3`] = `
 Error: both specification.path and specification.document are missing, should be path to the file or swagger document spec
 `
 
-exports[`test/static.js TAP specification validation check works > must match snapshot 4`] = `
+exports[`test/mode/static.js TAP specification validation check works > must match snapshot 4`] = `
 Error: specification.path is not a string
 `
 
-exports[`test/static.js TAP specification validation check works > must match snapshot 5`] = `
+exports[`test/mode/static.js TAP specification validation check works > must match snapshot 5`] = `
 Error: /hello/lionel.richie does not exist
 `
 
-exports[`test/static.js TAP specification validation check works > must match snapshot 6`] = `
+exports[`test/mode/static.js TAP specification validation check works > must match snapshot 6`] = `
 Error: specification.postProcessor should be a function
 `
 
-exports[`test/static.js TAP swagger route returns explicitly passed doc > must match snapshot 1`] = `
+exports[`test/mode/static.js TAP swagger route returns explicitly passed doc > must match snapshot 1`] = `
 {
   "info": {
     "title": "Test swagger",
