+++ test/decorator.js
@@ -1,0 +1,29 @@
+'use strict'
+
+const { test } = require('tap')
+const Fastify = require('fastify')
+const fastifySwagger = require('../index')
+
+test('fastify.swagger should exist', t => {
+  t.plan(2)
+  const fastify = Fastify()
+
+  fastify.register(fastifySwagger)
+
+  fastify.ready(err => {
+    t.error(err)
+    t.ok(fastify.swagger)
+  })
+})
+
+test('fastify.swaggerCSP should exist', t => {
+  t.plan(2)
+  const fastify = Fastify()
+
+  fastify.register(fastifySwagger)
+
+  fastify.ready(err => {
+    t.error(err)
+    t.ok(fastify.swaggerCSP)
+  })
+})
