+++ test/integration.js
@@ -1,7 +1,6 @@
 'use strict'
 
-const t = require('tap')
-const test = t.test
+const { test } = require('tap')
 const Fastify = require('fastify')
 const fastifySwagger = require('../index')
 const fastifyHelmet = require('fastify-helmet')
